pyproject.toml
[tool.poetry]
name = "sowba-gateway"
version = "0.1.0"
description = ""
authors = ["vangheem <vangheem@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
httpx = {extras = ["http2"], version = "^0.17.1"}
orjson = "^3.5.2"
uvicorn = "^0.13.4"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
pytest-asyncio = "^0.15.1"
async-asgi-testclient = "^1.4.6"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
api = 'sowba_gateway.commands:serve_command'

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
